# swagger/swagger.yaml
consumes:
- application/json
definitions:
  AddSamplesPayload:
    example:
      detail: sample1の詳細
      name: sample1のタイトル
    properties:
      detail:
        description: detail of sample
        example: sample1の詳細
        type: string
      name:
        description: name of sample
        example: sample1のタイトル
        type: string
    required:
    - name
    - detail
    title: AddSamplesPayload
    type: object
  Auth:
    description: Auth result (default view)
    example:
      token:
        expired_at: "2019-01-31T00:00:00+09:00"
        token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
    properties:
      token:
        $ref: '#/definitions/Token'
    required:
    - token
    title: 'Mediatype identifier: application/vnd.auth+json; view=default'
    type: object
  LoginAuthPayload:
    example:
      email: sample@goa-sample.test.com
      password: test1234
    properties:
      email:
        description: name of sample
        example: sample@goa-sample.test.com
        type: string
      password:
        description: detail of sample
        example: test1234
        type: string
    required:
    - email
    - password
    title: LoginAuthPayload
    type: object
  Sample:
    description: sample detail (default view)
    example:
      created_at: "2019-01-31T00:00:00+09:00"
      detail: サンプル1の詳細
      id: 1
      name: サンプル1
      updated_at: "2019-01-31T12:30:50+09:00"
      user_id: 1
    properties:
      created_at:
        description: 作成日
        example: "2019-01-31T00:00:00+09:00"
        format: date-time
        type: string
      detail:
        description: 詳細
        example: サンプル1の詳細
        type: string
      id:
        description: sample id
        example: 1
        format: int64
        type: integer
      name:
        description: 名前
        example: サンプル1
        type: string
      updated_at:
        description: 更新日
        example: "2019-01-31T12:30:50+09:00"
        format: date-time
        type: string
      user_id:
        description: user id
        example: 1
        format: int64
        type: integer
    required:
    - id
    - user_id
    - name
    - detail
    - created_at
    - updated_at
    title: 'Mediatype identifier: application/vnd.sample+json; view=default'
    type: object
  Samples:
    description: sample list (default view)
    example:
      created_at: "2019-01-31T00:00:00+09:00"
      id: 1
      name: サンプル1
      updated_at: "2019-01-31T12:30:50+09:00"
    properties:
      created_at:
        description: 作成日
        example: "2019-01-31T00:00:00+09:00"
        format: date-time
        type: string
      id:
        description: id
        example: 1
        format: int64
        type: integer
      name:
        description: 名前
        example: サンプル1
        type: string
      updated_at:
        description: 更新日
        example: "2019-01-31T12:30:50+09:00"
        format: date-time
        type: string
    required:
    - id
    - name
    - created_at
    - updated_at
    title: 'Mediatype identifier: application/vnd.samples+json; view=default'
    type: object
  SamplesCollection:
    description: SamplesCollection is the media type for an array of Samples (default
      view)
    example:
    - created_at: "2019-01-31T00:00:00+09:00"
      id: 1
      name: サンプル1
      updated_at: "2019-01-31T12:30:50+09:00"
    - created_at: "2019-01-31T00:00:00+09:00"
      id: 1
      name: サンプル1
      updated_at: "2019-01-31T12:30:50+09:00"
    - created_at: "2019-01-31T00:00:00+09:00"
      id: 1
      name: サンプル1
      updated_at: "2019-01-31T12:30:50+09:00"
    items:
      $ref: '#/definitions/Samples'
    title: 'Mediatype identifier: application/vnd.samples+json; type=collection; view=default'
    type: array
  Token:
    description: token (default view)
    example:
      expired_at: "2019-01-31T00:00:00+09:00"
      token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
    properties:
      expired_at:
        description: 期限
        example: "2019-01-31T00:00:00+09:00"
        format: date-time
        type: string
      token:
        description: token value
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        type: string
    required:
    - token
    - expired_at
    title: 'Mediatype identifier: application/vnd.token+json; view=default'
    type: object
  UpdateSamplesPayload:
    example:
      detail: sample1の詳細
      name: sample1のタイトル
    properties:
      detail:
        description: detail of sample
        example: sample1の詳細
        type: string
      name:
        description: name of sample
        example: sample1のタイトル
        type: string
    required:
    - name
    - detail
    title: UpdateSamplesPayload
    type: object
  error:
    description: Error response media type (default view)
    example:
      code: invalid_value
      detail: Value of ID must be an integer
      id: 3F1FKVRR
      meta:
        timestamp: 1458609066
      status: "400"
    properties:
      code:
        description: an application-specific error code, expressed as a string value.
        example: invalid_value
        type: string
      detail:
        description: a human-readable explanation specific to this occurrence of the
          problem.
        example: Value of ID must be an integer
        type: string
      id:
        description: a unique identifier for this particular occurrence of the problem.
        example: 3F1FKVRR
        type: string
      meta:
        additionalProperties: true
        description: a meta object containing non-standard meta-information about
          the error.
        example:
          timestamp: 1458609066
        type: object
      status:
        description: the HTTP status code applicable to this problem, expressed as
          a string value.
        example: "400"
        type: string
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
host: localhost:8080
info:
  description: A simple goa service
  title: The Sample API
  version: v1
paths:
  /api/v1/login:
    post:
      operationId: auth#login
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/LoginAuthPayload'
      produces:
      - application/vnd.auth+json
      - application/vnd.goa.error
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Auth'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
      schemes:
      - http
      - https
      summary: login auth
      tags:
      - auth
  /api/v1/samples/:
    get:
      description: |-
        複数

        Required security scopes:
          * `api:access`
      operationId: samples#list
      produces:
      - application/vnd.goa.error
      - application/vnd.samples+json; type=collection
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SamplesCollection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
      schemes:
      - http
      - https
      security:
      - jwt:
        - api:access
      summary: list samples
      tags:
      - samples
    post:
      description: |-
        追加

        Required security scopes:
          * `api:access`
      operationId: samples#add
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/AddSamplesPayload'
      produces:
      - application/vnd.goa.error
      - application/vnd.sample+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Sample'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
      schemes:
      - http
      - https
      security:
      - jwt:
        - api:access
      summary: add samples
      tags:
      - samples
  /api/v1/samples/{id}:
    delete:
      description: |-
        削除

        Required security scopes:
          * `api:access`
      operationId: samples#delete
      parameters:
      - description: sample id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/vnd.goa.error
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
      schemes:
      - http
      - https
      security:
      - jwt:
        - api:access
      summary: delete samples
      tags:
      - samples
    get:
      description: |-
        単数

        Required security scopes:
          * `api:access`
      operationId: samples#show
      parameters:
      - description: sample id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/vnd.goa.error
      - application/vnd.sample+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Sample'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
      schemes:
      - http
      - https
      security:
      - jwt:
        - api:access
      summary: show samples
      tags:
      - samples
    patch:
      description: |-
        更新

        Required security scopes:
          * `api:access`
      operationId: samples#update
      parameters:
      - description: sample id
        in: path
        name: id
        required: true
        type: integer
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/UpdateSamplesPayload'
      produces:
      - application/vnd.goa.error
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
      schemes:
      - http
      - https
      security:
      - jwt:
        - api:access
      summary: update samples
      tags:
      - samples
  /swagger-ui/{filepath}:
    get:
      operationId: swagger#/swagger-ui/*filepath
      parameters:
      - description: Relative file path
        in: path
        name: filepath
        required: true
        type: string
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
        "404":
          description: File not found
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      - https
      summary: Download public/swagger-ui/dist
  /swagger.json:
    get:
      operationId: swagger#/swagger.json
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
      schemes:
      - http
      - https
      summary: Download swagger/swagger.json
  /swagger.yaml:
    get:
      operationId: swagger#/swagger.yaml
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
      schemes:
      - http
      - https
      summary: Download swagger/swagger.yaml
produces:
- application/json
responses:
  NoContent:
    description: No Content
  NotFound:
    description: Not Found
schemes:
- http
- https
securityDefinitions:
  jwt:
    description: |2-


      **Security Scopes**:
        * `api:access`: API access
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"